+++ lib/retryable-monitor/index.ts
@@ -8,7 +8,6 @@ import {
   addCustomNetwork,
   L1TransactionReceipt as ParentChainTxReceipt,
   L1ToL2MessageStatus as ParentToChildMessageStatus,
-  L2Network as ParentNetwork,
   getL2Network,
   L1ToL2MessageStatus,
   L1ToL2MessageReader,
@@ -27,30 +26,15 @@ import {
   ARB_MINIMUM_BLOCK_TIME_IN_SECONDS,
   SEVEN_DAYS_IN_SECONDS,
 } from '@arbitrum/sdk/dist/lib/dataEntities/constants'
+import { getExplorerUrlPrefixes, reportFailedTicket } from './reportRetryables'
+import { reportRetryableErrorToSlack } from './reportRetryableErrorToSlack'
 import {
   ChildChainTicketReport,
+  ChildNetwork,
+  FindRetryablesOptions,
   ParentChainTicketReport,
   TokenDepositData,
-  getExplorerUrlPrefixes,
-  reportFailedTicket,
-} from './reportRetryables'
-import { reportRetryableErrorToSlack } from './reportRetryableErrorToSlack'
-
-// Interface defining additional properties for ChildNetwork
-export interface ChildNetwork extends ParentNetwork {
-  parentRpcUrl: string
-  orbitRpcUrl: string
-  parentExplorerUrl: string
-}
-
-// Type for options passed to findRetryables function
-type findRetryablesOptions = {
-  fromBlock: number
-  toBlock: number
-  continuous: boolean
-  configPath: string
-  enableAlerting: boolean
-}
+} from './types'
 
 // Path for the log file
 const logFilePath = 'logfile.log'
@@ -113,7 +97,7 @@ const checkNetworkAlreadyExistsInSdk = async (networkId: number) => {
 }
 
 // Parsing command line arguments using yargs
-const options: findRetryablesOptions = yargs(process.argv.slice(2))
+const options: FindRetryablesOptions = yargs(process.argv.slice(2))
   .options({
     fromBlock: { type: 'number', default: 0 },
     toBlock: { type: 'number', default: 0 },
@@ -122,12 +106,12 @@ const options: findRetryablesOptions = yargs(process.argv.slice(2))
     enableAlerting: { type: 'boolean', default: false },
   })
   .strict()
-  .parseSync() as findRetryablesOptions
+  .parseSync() as FindRetryablesOptions
 
 // Function to process a child chain and check for retryable transactions
 const processChildChain = async (
   childChain: ChildNetwork,
-  options: findRetryablesOptions
+  options: FindRetryablesOptions
 ) => {
   console.log('----------------------------------------------------------')
   console.log(`Running for Chain: ${childChain.name}`)
